minor changes after review

+++ clea/clea-venue-consumer/src/main/java/fr/gouv/clea/consumer/configuration/VenueConsumerConfiguration.java
@@ -9,5 +9,5 @@ import lombok.Data;
 @Configuration
 @ConfigurationProperties(prefix = "clea.conf")
 public class VenueConsumerConfiguration {
-    private int durationUnitInSeconds;
+    private long durationUnitInSeconds;
 }
